Add an internal folder
