+++ internal/cmd/inbox_delete.go
