+++ internal/cmd/inbox_test.go
