+++ internal/cmd/version_test.go
