+++ internal/cmd/view.go
