+++ internal/inbox/inbox_empty.html
