+++ internal/inbox/mail.html
