+++ internal/inbox/main_page.html
