+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/antham/yogo/cmd"
+import "github.com/antham/yogo/internal/cmd"
 
 func main() {
 	cmd.Execute()
